update: simplify and refactor codebase
This commit drastically simplifies the `nvim-lua-gf` codebase, and now
it only requires a fraction of the code for the similar effect as older
commits.

I've been reading a lot on the Vim options `include`, `includeexpr`,
`path`, and `suffixesadd`. It turns out that we only need to perform a
simple substitution on the filename to create a real file path, and adding
the Neovim runtime paths directly to `path` will perform an automatic
search through all Lua files in the Neovim runtime paths!